Moved older entry detail page in old folder and these pages will render on older routes.
